+++ demosplan/DemosPlanCoreBundle/Controller/User/DemosPlanDepartmentAPIController.php
@@ -12,7 +12,7 @@ namespace demosplan\DemosPlanCoreBundle\Controller\User;
 
 use DemosEurope\DemosplanAddon\Controller\APIController;
 use DemosEurope\DemosplanAddon\Response\APIResponse;
-use demosplan\DemosPlanCoreBundle\Annotation\DplanPermissions;
+use demosplan\DemosPlanCoreBundle\Attribute\DplanPermissions;
 use demosplan\DemosPlanCoreBundle\Logic\User\OrgaHandler;
 use demosplan\DemosPlanCoreBundle\ResourceTypes\DepartmentResourceType;
 use demosplan\DemosPlanCoreBundle\Services\ApiResourceService;
@@ -21,10 +21,9 @@ use Symfony\Component\Routing\Attribute\Route;
 class DemosPlanDepartmentAPIController extends APIController
 {
     /**
-     * @DplanPermissions("area_manage_users")
-     *
      * @param string $organisationId
      */
+    #[DplanPermissions('area_manage_users')]
     #[Route(path: '/api/1.0/{organisationId}/department', methods: ['GET'], name: 'dplan_api_department_list', options: ['expose' => true])]
     public function list(ApiResourceService $apiResourceService, OrgaHandler $orgaHandler, $organisationId): APIResponse
     {