+++ demosplan/DemosPlanCoreBundle/Logic/ViewRenderer.php
@@ -124,7 +124,7 @@ class ViewRenderer
      *
      * @return RedirectResponse|Response Response
      *
-     * @deprecated use DplanPermissions({"permission"}) Annotation on controllers instead
+     * @deprecated use DplanPermissions({"permission"}) Attribute on controllers instead
      *     try/catch can be omitted, this error handling has moved to ExceptionListener
      *     and CheckPermissionListener
      *