+++ CHANGELOG.md
@@ -5,6 +5,16 @@
 - **Patch Version**: Incremented for bug fixes.
 
 ## UNRELEASED
+
+- Add tag-based filtering for segments of Statement exports (DOCX, XLSX, ZIP)
+  that filters segments within statements
+  by tag ID, tag title, tag topic ID, or tag topic title.
+  Statements without matching segments are excluded from export.
+  No applied tag-filter will still export all statements unchanged
+  
+- Add similar submitters to the Submitter List
+- Add segment text to the boilerplate modal, if the segment and its text is available
+
 ## v4.25.0 (2025-11-06)
 ## v4.24.0 (2025-11-06)
 - Detect Company Department from OzgKeycloak token and assign it to user
@@ -27,10 +37,12 @@
 ### Features
 - Add possibility to delete custom fields and their options
 
+## v4.15.3 (2025-12-02)
 ## v4.15.2 (2025-10-24)
 - fix zip download for older uploads
 
 ## v4.15.0 (2025-09-15)
+## v4.14.2 (2025-12-02)
 ## v4.14.0 (2025-09-15)
 - Add html paragraph import from odt files